+++ examples/example-bckey-internal.c
@@ -97,7 +97,7 @@
 
   printf("key ---> %s (%s)\n",
                  ak_buffer_get_str( &bkey->key.oid->name ), ak_buffer_get_str(& bkey->key.oid->id ));
-  printf("  number: %s\n", ak_buffer_get_str( &bkey->key.number ));
+  printf("  number: %s\n", str = ak_buffer_to_hexstr( &bkey->key.number )); free( str );
   printf("   flags: %016llx\n", bkey->key.flags );
   printf(" counter: %016llx (%llu)\n", bkey->key.resource.counter, bkey->key.resource.counter );
   printf("     key: %s\n", str = ak_ptr_to_hexstr( bkey->key.key.data, 32, ak_false )); free( str );