+++ Changelog.md
@@ -1,7 +1,8 @@
 Changes in version 0.5.7
 -------------------------------------
-  * Changed internal structure of libakrypt options.
-    Now we can read their values from userspace (see libakrypt.h).
+  * Restored block cipher Magma and ecb, xcrypt encryption modes
+  * Changed internal structure of libakrypt options
+    Now we can read their values from userspace (see libakrypt.h)
 
 
 Changes in version 0.5.6 (29.11.2017)
@@ -19,7 +20,7 @@ Changes in version 0.5.4 (10.10.2017)
 -------------------------------------
 
    * Restored arithmetic in prime fields with Montgomery improovments
-   * Restored operations on short Weierstrass elliptic curves.
+   * Restored operations on short Weierstrass elliptic curves
    * Added 256 & 512 bits parameters for useful Weierstrass curves
    * Improoved handbook